requirements.txt
beautifulsoup4
requests
repoze.lru